fix header image not loading & sign-up button wrap
Once I added the webpage to Github pages I had an issue where the header
image would not load and the sign up button text in .section-4 was wrap-
ping when I didn't want it to. The issues did not happen when I was
viewing the page through VScodes live server. Replaced the image that
was not loading with one that was.